+++ qtfred/src/mission/dialogs/EventEditorDialogModel.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "AbstractDialogModel.h"
+
 namespace fso {
 namespace fred {
 namespace dialogs {
@@ -8,7 +9,7 @@ namespace dialogs {
 class EventEditorDialogModel: public AbstractDialogModel {
  Q_OBJECT
  public:
-	EventEditorDialogModel(QObject* parent, Editor* editor, FredRenderer* renderer);
+	EventEditorDialogModel(QObject* parent, Editor* editor, EditorViewport* viewport);
 
 	void applyChanges();
 	void discardChanges();
