+++ qtfred/src/mission/management.cpp
@@ -17,6 +17,8 @@
 #include <starfield/starfield.h>
 #include <sound/audiostr.h>
 
+#include <clocale>
+
 extern int Xstr_inited;
 
 extern void allocate_mission_text(size_t size);
@@ -100,33 +102,6 @@ void shutdown() {
 #ifndef NDEBUG
 	outwnd_close();
 #endif
-}
-
-
-const char* object_name(int obj) {
-	static char text[80];
-	waypoint_list *wp_list;
-	int waypoint_num;
-
-	if (!query_valid_object(obj))
-		return "*none*";
-
-	switch (Objects[obj].type) {
-	case OBJ_SHIP:
-	case OBJ_START:
-		return Ships[Objects[obj].instance].ship_name;
-
-	case OBJ_WAYPOINT:
-		wp_list = find_waypoint_list_with_instance(Objects[obj].instance, &waypoint_num);
-		Assert(wp_list != NULL);
-		sprintf(text, "%s:%d", wp_list->get_name(), waypoint_num + 1);
-		return text;
-
-	case OBJ_POINT:
-		return "Briefing icon";
-	}
-
-	return "*unknown*";
 }
