+++ qtfred/src/ui/FredView.cpp
@@ -19,6 +19,7 @@
 #include "widgets/ColorComboBox.h"
 
 #include "util.h"
+#include "mission/object.h"
 
 namespace {
 
@@ -69,14 +70,14 @@ FredView::FredView(QWidget* parent) : QMainWindow(parent), ui(new Ui::FredView()
 FredView::~FredView() {
 }
 
-void FredView::setEditor(Editor* editor, FredRenderer* renderer) {
+void FredView::setEditor(Editor* editor, EditorViewport* viewport) {
 	Assertion(fred == nullptr, "Resetting the editor is currently not supported!");
-	Assertion(_renderer == nullptr, "Resetting the renderer is currently not supported!");
+	Assertion(_viewport == nullptr, "Resetting the viewport is currently not supported!");
 
 	fred = editor;
-	_renderer = renderer;
+	_viewport = viewport;
 
-	ui->centralWidget->setEditor(editor, renderer);
+	ui->centralWidget->setEditor(editor, _viewport);
 
 	connect(fred, &Editor::missionLoaded, this, &FredView::on_mission_loaded);
 
@@ -195,29 +196,29 @@ void FredView::recentFileOpened() {
 	loadMissionFile(path);
 }
 void FredView::syncViewOptions() {
-	connectActionToViewSetting(ui->actionShow_Ships, &_renderer->view.Show_ships);
-	connectActionToViewSetting(ui->actionShow_Player_Starts, &_renderer->view.Show_starts);
-	connectActionToViewSetting(ui->actionShow_Waypoints, &_renderer->view.Show_waypoints);
+	connectActionToViewSetting(ui->actionShow_Ships, &_viewport->view.Show_ships);
+	connectActionToViewSetting(ui->actionShow_Player_Starts, &_viewport->view.Show_starts);
+	connectActionToViewSetting(ui->actionShow_Waypoints, &_viewport->view.Show_waypoints);
 
 	// TODO: Dynamically handle the Show teams option
 
-	connectActionToViewSetting(ui->actionShow_Ship_Models, &_renderer->view.Show_ship_models);
-	connectActionToViewSetting(ui->actionShow_Outlines, &_renderer->view.Show_outlines);
-	connectActionToViewSetting(ui->actionShow_Ship_Info, &_renderer->view.Show_ship_info);
-	connectActionToViewSetting(ui->actionShow_Coordinates, &_renderer->view.Show_coordinates);
-	connectActionToViewSetting(ui->actionShow_Grid_Positions, &_renderer->view.Show_grid_positions);
-	connectActionToViewSetting(ui->actionShow_Distances, &_renderer->view.Show_distances);
-	connectActionToViewSetting(ui->actionShow_Model_Paths, &_renderer->view.Show_paths_fred);
-	connectActionToViewSetting(ui->actionShow_Model_Dock_Points, &_renderer->view.Show_dock_points);
+	connectActionToViewSetting(ui->actionShow_Ship_Models, &_viewport->view.Show_ship_models);
+	connectActionToViewSetting(ui->actionShow_Outlines, &_viewport->view.Show_outlines);
+	connectActionToViewSetting(ui->actionShow_Ship_Info, &_viewport->view.Show_ship_info);
+	connectActionToViewSetting(ui->actionShow_Coordinates, &_viewport->view.Show_coordinates);
+	connectActionToViewSetting(ui->actionShow_Grid_Positions, &_viewport->view.Show_grid_positions);
+	connectActionToViewSetting(ui->actionShow_Distances, &_viewport->view.Show_distances);
+	connectActionToViewSetting(ui->actionShow_Model_Paths, &_viewport->view.Show_paths_fred);
+	connectActionToViewSetting(ui->actionShow_Model_Dock_Points, &_viewport->view.Show_dock_points);
 
-	connectActionToViewSetting(ui->actionShow_Grid, &_renderer->view.Show_grid);
-	connectActionToViewSetting(ui->actionShow_Horizon, &_renderer->view.Show_horizon);
+	connectActionToViewSetting(ui->actionShow_Grid, &_viewport->view.Show_grid);
+	connectActionToViewSetting(ui->actionShow_Horizon, &_viewport->view.Show_horizon);
 	connectActionToViewSetting(ui->actionDouble_Fine_Gridlines, &double_fine_gridlines);
-	connectActionToViewSetting(ui->actionAnti_Aliased_Gridlines, &_renderer->view.Aa_gridlines);
-	connectActionToViewSetting(ui->actionShow_3D_Compass, &_renderer->view.Show_compass);
-	connectActionToViewSetting(ui->actionShow_Background, &_renderer->view.Show_stars);
+	connectActionToViewSetting(ui->actionAnti_Aliased_Gridlines, &_viewport->view.Aa_gridlines);
+	connectActionToViewSetting(ui->actionShow_3D_Compass, &_viewport->view.Show_compass);
+	connectActionToViewSetting(ui->actionShow_Background, &_viewport->view.Show_stars);
 
-	connectActionToViewSetting(ui->actionLighting_from_Suns, &_renderer->view.Lighting_on);
+	connectActionToViewSetting(ui->actionLighting_from_Suns, &_viewport->view.Lighting_on);
 }
 void FredView::initializeStatusBar() {
 	_statusBarViewmode = new QLabel();
@@ -227,15 +228,15 @@ void FredView::initializeStatusBar() {
 	statusBar()->addPermanentWidget(_statusBarUnitsLabel);
 }
 void FredView::updateUI() {
-	if (!_renderer) {
-		// The following code requires a valid renderer
+	if (!_viewport) {
+		// The following code requires a valid viewport
 		return;
 	}
 
-	_statusBarUnitsLabel->setText(tr("Units = %1 Meters").arg(_renderer->The_grid->square_size));
+	_statusBarUnitsLabel->setText(tr("Units = %1 Meters").arg(_viewport->The_grid->square_size));
 
-	if (_renderer->viewpoint == 1) {
-		_statusBarViewmode->setText(tr("Viewpoint: %1").arg(object_name(_renderer->view_obj)));
+	if (_viewport->viewpoint == 1) {
+		_statusBarViewmode->setText(tr("Viewpoint: %1").arg(object_name(_viewport->view_obj)));
 	} else {
 		_statusBarViewmode->setText(tr("Viewpoint: Camera"));
 	}
@@ -257,7 +258,7 @@ void FredView::connectActionToViewSetting(QAction* option, bool* destination) {
 		*destination = value;
 
 		// View settings have changed so we need to update the window
-		_renderer->scheduleUpdate();
+		_viewport->needsUpdate();
 	});
 }
 
@@ -293,58 +294,58 @@ void FredView::initializePopupMenus() {
 
 void FredView::onUpdateConstrains() {
 	ui->actionConstrainX->setChecked(
-		_renderer->Constraint.xyz.x && !_renderer->Constraint.xyz.y && !_renderer->Constraint.xyz.z);
+		_viewport->Constraint.xyz.x && !_viewport->Constraint.xyz.y && !_viewport->Constraint.xyz.z);
 	ui->actionConstrainY->setChecked(
-		!_renderer->Constraint.xyz.x && _renderer->Constraint.xyz.y && !_renderer->Constraint.xyz.z);
+		!_viewport->Constraint.xyz.x && _viewport->Constraint.xyz.y && !_viewport->Constraint.xyz.z);
 	ui->actionConstrainZ->setChecked(
-		!_renderer->Constraint.xyz.x && !_renderer->Constraint.xyz.y && _renderer->Constraint.xyz.z);
+		!_viewport->Constraint.xyz.x && !_viewport->Constraint.xyz.y && _viewport->Constraint.xyz.z);
 	ui->actionConstrainXZ->setChecked(
-		_renderer->Constraint.xyz.x && !_renderer->Constraint.xyz.y && _renderer->Constraint.xyz.z);
+		_viewport->Constraint.xyz.x && !_viewport->Constraint.xyz.y && _viewport->Constraint.xyz.z);
 	ui->actionConstrainXY->setChecked(
-		_renderer->Constraint.xyz.x && _renderer->Constraint.xyz.y && !_renderer->Constraint.xyz.z);
+		_viewport->Constraint.xyz.x && _viewport->Constraint.xyz.y && !_viewport->Constraint.xyz.z);
 	ui->actionConstrainYZ->setChecked(
-		!_renderer->Constraint.xyz.x && _renderer->Constraint.xyz.y && _renderer->Constraint.xyz.z);
+		!_viewport->Constraint.xyz.x && _viewport->Constraint.xyz.y && _viewport->Constraint.xyz.z);
 }
 void FredView::on_actionConstrainX_triggered(bool enabled) {
 	if (enabled) {
-		vm_vec_make(&_renderer->Constraint, 1.0f, 0.0f, 0.0f);
-		vm_vec_make(&_renderer->Anticonstraint, 0.0f, 1.0f, 1.0f);
-		_renderer->Single_axis_constraint = true;
+		vm_vec_make(&_viewport->Constraint, 1.0f, 0.0f, 0.0f);
+		vm_vec_make(&_viewport->Anticonstraint, 0.0f, 1.0f, 1.0f);
+		_viewport->Single_axis_constraint = true;
 	}
 }
 void FredView::on_actionConstrainY_triggered(bool enabled) {
 	if (enabled) {
-		vm_vec_make(&_renderer->Constraint, 0.0f, 1.0f, 0.0f);
-		vm_vec_make(&_renderer->Anticonstraint, 1.0f, 0.0f, 1.0f);
-		_renderer->Single_axis_constraint = true;
+		vm_vec_make(&_viewport->Constraint, 0.0f, 1.0f, 0.0f);
+		vm_vec_make(&_viewport->Anticonstraint, 1.0f, 0.0f, 1.0f);
+		_viewport->Single_axis_constraint = true;
 	}
 }
 void FredView::on_actionConstrainZ_triggered(bool enabled) {
 	if (enabled) {
-		vm_vec_make(&_renderer->Constraint, 0.0f, 0.0f, 1.0f);
-		vm_vec_make(&_renderer->Anticonstraint, 1.0f, 1.0f, 0.0f);
-		_renderer->Single_axis_constraint = true;
+		vm_vec_make(&_viewport->Constraint, 0.0f, 0.0f, 1.0f);
+		vm_vec_make(&_viewport->Anticonstraint, 1.0f, 1.0f, 0.0f);
+		_viewport->Single_axis_constraint = true;
 	}
 }
 void FredView::on_actionConstrainXZ_triggered(bool enabled) {
 	if (enabled) {
-		vm_vec_make(&_renderer->Constraint, 1.0f, 0.0f, 1.0f);
-		vm_vec_make(&_renderer->Anticonstraint, 0.0f, 1.0f, 0.0f);
-		_renderer->Single_axis_constraint = false;
+		vm_vec_make(&_viewport->Constraint, 1.0f, 0.0f, 1.0f);
+		vm_vec_make(&_viewport->Anticonstraint, 0.0f, 1.0f, 0.0f);
+		_viewport->Single_axis_constraint = false;
 	}
 }
 void FredView::on_actionConstrainXY_triggered(bool enabled) {
 	if (enabled) {
-		vm_vec_make(&_renderer->Constraint, 1.0f, 1.0f, 0.0f);
-		vm_vec_make(&_renderer->Anticonstraint, 0.0f, 0.0f, 1.0f);
-		_renderer->Single_axis_constraint = false;
+		vm_vec_make(&_viewport->Constraint, 1.0f, 1.0f, 0.0f);
+		vm_vec_make(&_viewport->Anticonstraint, 0.0f, 0.0f, 1.0f);
+		_viewport->Single_axis_constraint = false;
 	}
 }
 void FredView::on_actionConstrainYZ_triggered(bool enabled) {
 	if (enabled) {
-		vm_vec_make(&_renderer->Constraint, 0.0f, 1.0f, 1.0f);
-		vm_vec_make(&_renderer->Anticonstraint, 1.0f, 0.0f, 0.0f);
-		_renderer->Single_axis_constraint = false;
+		vm_vec_make(&_viewport->Constraint, 0.0f, 1.0f, 1.0f);
+		vm_vec_make(&_viewport->Anticonstraint, 1.0f, 0.0f, 0.0f);
+		_viewport->Single_axis_constraint = false;
 	}
 }
 RenderWidget* FredView::getRenderWidget() {
@@ -352,25 +353,25 @@ RenderWidget* FredView::getRenderWidget() {
 }
 void FredView::on_actionSelect_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->Editing_mode = CursorMode::Selecting;
+		_viewport->Editing_mode = CursorMode::Selecting;
 	}
 }
 void FredView::on_actionSelectMove_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->Editing_mode = CursorMode::Moving;
+		_viewport->Editing_mode = CursorMode::Moving;
 	}
 }
 void FredView::on_actionSelectRotate_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->Editing_mode = CursorMode::Rotating;
+		_viewport->Editing_mode = CursorMode::Rotating;
 	}
 }
 void FredView::onUpdateEditingMode() {
-	ui->actionSelect->setChecked(_renderer->Editing_mode == CursorMode::Selecting);
-	ui->actionSelectMove->setChecked(_renderer->Editing_mode == CursorMode::Moving);
-	ui->actionSelectRotate->setChecked(_renderer->Editing_mode == CursorMode::Rotating);
+	ui->actionSelect->setChecked(_viewport->Editing_mode == CursorMode::Selecting);
+	ui->actionSelectMove->setChecked(_viewport->Editing_mode == CursorMode::Moving);
+	ui->actionSelectRotate->setChecked(_viewport->Editing_mode == CursorMode::Rotating);
 
-	ui->centralWidget->setCursorMode(_renderer->Editing_mode);
+	ui->centralWidget->setCursorMode(_viewport->Editing_mode);
 }
 bool FredView::event(QEvent* event) {
 	switch (event->type()) {
@@ -387,12 +388,12 @@ bool FredView::event(QEvent* event) {
 void FredView::windowActivated() {
 	key_got_focus();
 
-	_renderer->Cursor_over = -1;
+	_viewport->Cursor_over = -1;
 }
 void FredView::windowDeactivated() {
 	key_lost_focus();
 
-	_renderer->Cursor_over = -1;
+	_viewport->Cursor_over = -1;
 }
 void FredView::on_actionHide_Marked_Objects_triggered(bool enabled) {
 	fred->hideMarkedObjects();
@@ -401,129 +402,129 @@ void FredView::on_actionShow_All_Hidden_Objects_triggered(bool enabled) {
 	fred->showHiddenObjects();
 }
 void FredView::onUpdateViewSpeeds() {
-	ui->actionx1->setChecked(_renderer->physics_speed == 1);
-	ui->actionx2->setChecked(_renderer->physics_speed == 2);
-	ui->actionx3->setChecked(_renderer->physics_speed == 3);
-	ui->actionx5->setChecked(_renderer->physics_speed == 5);
-	ui->actionx8->setChecked(_renderer->physics_speed == 8);
-	ui->actionx10->setChecked(_renderer->physics_speed == 10);
-	ui->actionx50->setChecked(_renderer->physics_speed == 50);
-	ui->actionx100->setChecked(_renderer->physics_speed == 100);
+	ui->actionx1->setChecked(_viewport->physics_speed == 1);
+	ui->actionx2->setChecked(_viewport->physics_speed == 2);
+	ui->actionx3->setChecked(_viewport->physics_speed == 3);
+	ui->actionx5->setChecked(_viewport->physics_speed == 5);
+	ui->actionx8->setChecked(_viewport->physics_speed == 8);
+	ui->actionx10->setChecked(_viewport->physics_speed == 10);
+	ui->actionx50->setChecked(_viewport->physics_speed == 50);
+	ui->actionx100->setChecked(_viewport->physics_speed == 100);
 
-	ui->actionRotx1->setChecked(_renderer->physics_rot == 2);
-	ui->actionRotx5->setChecked(_renderer->physics_rot == 10);
-	ui->actionRotx12->setChecked(_renderer->physics_rot == 25);
-	ui->actionRotx25->setChecked(_renderer->physics_rot == 50);
-	ui->actionRotx50->setChecked(_renderer->physics_rot == 100);
+	ui->actionRotx1->setChecked(_viewport->physics_rot == 2);
+	ui->actionRotx5->setChecked(_viewport->physics_rot == 10);
+	ui->actionRotx12->setChecked(_viewport->physics_rot == 25);
+	ui->actionRotx25->setChecked(_viewport->physics_rot == 50);
+	ui->actionRotx50->setChecked(_viewport->physics_rot == 100);
 }
 void FredView::on_actionx1_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 1;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 1;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx2_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 2;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 2;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx3_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 3;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 3;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx5_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 5;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 5;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx8_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 8;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 8;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx10_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 10;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 10;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx50_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 50;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 50;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionx100_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_speed = 100;
-		_renderer->resetViewPhysics();
+		_viewport->physics_speed = 100;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionRotx1_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_rot = 2;
-		_renderer->resetViewPhysics();
+		_viewport->physics_rot = 2;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionRotx5_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_rot = 10;
-		_renderer->resetViewPhysics();
+		_viewport->physics_rot = 10;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionRotx12_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_rot = 25;
-		_renderer->resetViewPhysics();
+		_viewport->physics_rot = 25;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionRotx25_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_rot = 50;
-		_renderer->resetViewPhysics();
+		_viewport->physics_rot = 50;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::on_actionRotx50_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->physics_rot = 100;
-		_renderer->resetViewPhysics();
+		_viewport->physics_rot = 100;
+		_viewport->resetViewPhysics();
 	}
 }
 void FredView::onUpdateCameraControlActions() {
-	ui->actionCamera->setChecked(_renderer->viewpoint == 0);
-	ui->actionCurrent_Ship->setChecked(_renderer->viewpoint == 1);
+	ui->actionCamera->setChecked(_viewport->viewpoint == 0);
+	ui->actionCurrent_Ship->setChecked(_viewport->viewpoint == 1);
 
-	_controlModeCamera->setChecked(_renderer->Control_mode == 0);
-	_controlModeCurrentShip->setChecked(_renderer->Control_mode == 1);
+	_controlModeCamera->setChecked(_viewport->Control_mode == 0);
+	_controlModeCurrentShip->setChecked(_viewport->Control_mode == 1);
 }
 void FredView::on_actionCamera_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->viewpoint = 0;
+		_viewport->viewpoint = 0;
 
-		_renderer->scheduleUpdate();
+		_viewport->needsUpdate();
 	}
 }
 void FredView::on_actionCurrent_Ship_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->viewpoint = 1;
-		_renderer->view_obj = fred->getCurrentObject();
+		_viewport->viewpoint = 1;
+		_viewport->view_obj = fred->getCurrentObject();
 
-		_renderer->scheduleUpdate();
+		_viewport->needsUpdate();
 	}
 }
 void FredView::on_actionControlModeCamera_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->Control_mode = 0;
+		_viewport->Control_mode = 0;
 	}
 }
 void FredView::on_actionControlModeCurrentShip_triggered(bool enabled) {
 	if (enabled) {
-		_renderer->Control_mode = 1;
+		_viewport->Control_mode = 1;
 	}
 }
 
@@ -548,7 +549,7 @@ void FredView::keyReleaseEvent(QKeyEvent* event) {
 	_inKeyReleaseHandler = false;
 }
 void FredView::on_actionEvents_triggered(bool) {
-	auto eventEditor = new dialogs::EventEditorDialog(this, fred, _renderer);
+	auto eventEditor = new dialogs::EventEditorDialog(this, fred, _viewport);
 	eventEditor->show();
 }
