+++ qtfred/src/ui/dialogs/EventEditorDialog.cpp
@@ -8,8 +8,8 @@ namespace fso {
 namespace fred {
 namespace dialogs {
 
-EventEditorDialog::EventEditorDialog(QWidget* parent, Editor* editor, FredRenderer* renderer) :
-	QDialog(parent), ui(new Ui::EventEditorDialog()), _model(new EventEditorDialogModel(this, editor, renderer)) {
+EventEditorDialog::EventEditorDialog(QWidget* parent, Editor* editor, EditorViewport* viewport) :
+	QDialog(parent), ui(new Ui::EventEditorDialog()), _model(new EventEditorDialogModel(this, editor, viewport)) {
 	ui->setupUi(this);
 
 	connect(this, &QDialog::accepted, _model.get(), &EventEditorDialogModel::applyChanges);
